Change XProcSpecRunner API
Instead of returning all results at the end, provide a `Reporter'
interface with callbacks for notifications and for returning
intermediate results after each test. The default Reporter prints a text
based report to stdout. Deprecates the `TestLogger' interface.

The Reporter interface could be improved over time to allow for more
advanced reports. Currently the test results are much less detailed in
Java than in HTML. This is because each test is run externally with
XProcSpec and the results are obtained by parsing the Surefire
reports. These reports are currently not expressive enough; a lot of the
metadata and structure of the original test format is lost.

One idea to improve on this is to parse the "internal" XProcSpec report
format. To make this more robust the internal format should be made more
official.

Another idea to improve the integration is to capture XProcSpec's log
messages and make them available in Java through callbacks. A possible
solution would be to override `xproc/utils/log.xpl` with a custom
Calabash step.

An implementation of XProcSpec in Java would eliminate the need for
parsing XML reports completely, and it would also make the logging
integration easier.